create videocapture to handle videosetting
