+++ Tireless/View/Base.lproj/Home.storyboard
@@ -4,7 +4,6 @@
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="20020"/>
-        <capability name="Named colors" minToolsVersion="9.0"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="collection view cell content view" minToolsVersion="11.0"/>
@@ -82,19 +81,6 @@
                                 <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 <nil key="highlightedColor"/>
                             </label>
-                            <button hidden="YES" opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="x3I-UH-LHj">
-                                <rect key="frame" x="182" y="762" width="50" height="50"/>
-                                <color key="backgroundColor" name="themeYellow"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="50" id="IjT-ud-8HI"/>
-                                    <constraint firstAttribute="width" secondItem="x3I-UH-LHj" secondAttribute="height" id="ccg-sd-Afr"/>
-                                </constraints>
-                                <color key="tintColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
-                                <connections>
-                                    <action selector="recordTap:" destination="BYZ-38-t0r" eventType="touchUpInside" id="Fc5-2Q-zE6"/>
-                                </connections>
-                            </button>
                             <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="QXz-2S-P4b">
                                 <rect key="frame" x="20" y="64" width="50" height="50"/>
                                 <constraints>
@@ -116,7 +102,6 @@
                             <constraint firstItem="YmP-KB-PVm" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="80" id="5uO-nf-SZG"/>
                             <constraint firstItem="0am-LU-uPW" firstAttribute="centerY" secondItem="8bC-Xf-vdC" secondAttribute="centerY" id="6t9-AM-5q7"/>
                             <constraint firstItem="QXz-2S-P4b" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="20" id="85w-yA-PBC"/>
-                            <constraint firstItem="6Tk-OE-BBY" firstAttribute="bottom" secondItem="x3I-UH-LHj" secondAttribute="bottom" constant="50" id="Do6-b0-FLr"/>
                             <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="0am-LU-uPW" secondAttribute="trailing" constant="10" id="PRa-gZ-g03"/>
                             <constraint firstAttribute="trailing" secondItem="KWo-Ub-40G" secondAttribute="trailing" id="Uph-iV-67C"/>
                             <constraint firstAttribute="bottom" secondItem="KWo-Ub-40G" secondAttribute="bottom" id="Wne-fG-WQ4"/>
@@ -124,14 +109,12 @@
                             <constraint firstItem="QXz-2S-P4b" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="20" id="sfb-yQ-qGy"/>
                             <constraint firstItem="YmP-KB-PVm" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="50" id="t49-59-9S7"/>
                             <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="YmP-KB-PVm" secondAttribute="trailing" constant="50" id="vtE-67-tmZ"/>
-                            <constraint firstItem="x3I-UH-LHj" firstAttribute="centerX" secondItem="8bC-Xf-vdC" secondAttribute="centerX" id="zQc-wc-V9T"/>
                         </constraints>
                     </view>
                     <connections>
                         <outlet property="cameraPreView" destination="KWo-Ub-40G" id="lko-HD-8Dn"/>
                         <outlet property="countLabel" destination="0am-LU-uPW" id="7Gh-nM-yea"/>
                         <outlet property="inFrameLikeLiHoodLabel" destination="YmP-KB-PVm" id="HzO-CI-Bq3"/>
-                        <outlet property="recordButton" destination="x3I-UH-LHj" id="544-YY-8qT"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
