+++ Tireless/ViewModel/Plan/PlanDetailViewModel.swift
@@ -18,6 +18,10 @@ class PlanDetailViewModel {
                           finishTime: [],
                           uuid: "")
     
+//    init(plan: Plan) {
+//        self.plan = plan
+//    }
+    
     func setPlanData(name: String, times: String, days: String, createdTime: Int64, planGroup: Bool) {
         self.plan.planName = name
         self.plan.planTimes = times
